+++ bc_obps/compliance/api/_bccr/_accounts/_account_id/_compliance_report_versions/_compliance_report_version_id/compliance_units.py
@@ -5,7 +5,7 @@ from typing import Dict, Literal, Tuple
 from django.http import HttpRequest
 from common.permissions import authorize
 from compliance.schema.apply_compliance_units import ApplyComplianceUnitsIn, ApplyComplianceUnitsOut
-from compliance.service.apply_compliance_units_service import ApplyComplianceUnitsService
+from compliance.service.bc_carbon_registry.apply_compliance_units_service import ApplyComplianceUnitsService
 from compliance.service.bc_carbon_registry.schema import FifteenDigitString
 from service.error_service.custom_codes_4xx import custom_codes_4xx
 from registration.schema.generic import Message
