+++ bc_obps/compliance/tests/api/_bccr/_accounts/_account_id/_compliance_report_versions/_compliance_report_version_id/test_compliance_units.py
@@ -7,13 +7,9 @@ from compliance.dataclass import BCCRUnit, ComplianceUnitsPageData
 # Constants
 VALID_ACCOUNT_ID = "123456789012345"
 VALID_COMPLIANCE_REPORT_VERSION_ID = 1
-APPLY_COMPLIANCE_UNITS_SERVICE_PATH = (
-    "compliance.service.apply_compliance_units_service.ApplyComplianceUnitsService.get_apply_compliance_units_page_data"
-)
+APPLY_COMPLIANCE_UNITS_SERVICE_PATH = "compliance.service.bc_carbon_registry.apply_compliance_units_service.ApplyComplianceUnitsService.get_apply_compliance_units_page_data"
 PERMISSION_CHECK_PATH = "common.permissions.check_permission_for_role"
-APPLY_COMPLIANCE_UNITS_SERVICE_APPLY_PATH = (
-    "compliance.service.apply_compliance_units_service.ApplyComplianceUnitsService.apply_compliance_units"
-)
+APPLY_COMPLIANCE_UNITS_SERVICE_APPLY_PATH = "compliance.service.bc_carbon_registry.apply_compliance_units_service.ApplyComplianceUnitsService.apply_compliance_units"
 
 
 @override_settings(MIDDLEWARE=[])  # Disable middleware to prevent database queries
